requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements --dev
#

# Note: in pipenv 2020.x, "--dev" changed to emit both default and development
# requirements. To emit only development requirements, pass "--dev-only".

-i https://pypi.org/simple
absl-py==0.10.0
anyio==2.0.2; python_full_version >= '3.6.2'
argon2-cffi==20.1.0
astroid==2.4.2; python_version >= '3.5'
astunparse==1.6.3
async-generator==1.10; python_version >= '3.5'
attrs==20.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
babel==2.9.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
backcall==0.2.0
bleach==3.2.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cachetools==4.2.1; python_version ~= '3.5'
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
click==7.1.2
datasets==1.2.1
decorator==4.4.2
defusedxml==0.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dill==0.3.3; python_version >= '2.6' and python_version != '3.0'
entrypoints==0.3; python_version >= '2.7'
fastapi==0.63.0
filelock==3.0.12
flatbuffers==1.12
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
gast==0.3.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-auth-oauthlib==0.4.2; python_version >= '3.6'
google-auth==1.24.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-pasta==0.2.0
googleapis-common-protos==1.52.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
grpcio==1.32.0
h11==0.12.0; python_version >= '3.6'
h5py==2.10.0
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
importlib-metadata==3.4.0; python_version < '3.8' and python_version < '3.8'
importlib-resources==5.1.0; python_version < '3.9'
iniconfig==1.1.1
ipykernel==5.4.3; python_version >= '3.5'
ipython-genutils==0.2.0
ipython==7.19.0; python_version >= '3.3'
ipywidgets==7.6.3
isort==5.7.0; python_version >= '3.6' and python_version < '4.0'
jedi==0.18.0; python_version >= '3.6'
jinja2==2.11.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
joblib==1.0.0; python_version >= '3.6'
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.11; python_version >= '3.5'
jupyter-console==6.2.0; python_version >= '3.6'
jupyter-core==4.7.0; python_version >= '3.6'
jupyter-server==1.2.2; python_version >= '3.6'
jupyter==1.0.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.1.2; python_version >= '3.6'
jupyterlab-widgets==1.0.0; python_version >= '3.6'
jupyterlab==3.0.5
keras-preprocessing==1.1.2
lazy-object-proxy==1.4.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
markdown==3.3.3; python_version >= '3.6'
markupsafe==1.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
mccabe==0.6.1
mistune==0.8.4
multiprocess==0.70.11.1
nbclassic==0.2.6; python_version >= '3.6'
nbclient==0.5.1; python_version >= '3.6'
nbconvert==6.0.7; python_version >= '3.6'
nbformat==5.1.2; python_version >= '3.5'
nest-asyncio==1.4.3; python_version >= '3.5'
notebook==6.2.0; python_version >= '3.5'
numpy==1.19.5; python_version >= '3.6'
oauthlib==3.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
opt-einsum==3.3.0; python_version >= '3.5'
packaging==20.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pandas==1.2.1
pandocfilters==1.4.3
parso==0.8.1; python_version >= '3.6'
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
plotly==4.14.3
pluggy==0.13.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.14; python_full_version >= '3.6.1'
protobuf==3.14.0
ptyprocess==0.7.0; os_name != 'nt'
py==1.10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyarrow==3.0.0; python_version >= '3.6'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pydantic==1.7.3; python_version >= '3.6'
pygments==2.7.4; python_version >= '3.5'
pylint==2.6.0
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
pyrsistent==0.17.3; python_version >= '3.5'
pytest==6.2.2
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pytz==2020.5
pyzmq==21.0.2; python_version >= '3.6'
qtconsole==5.0.2; python_version >= '3.6'
qtpy==1.9.0
regex==2020.11.13
requests-oauthlib==1.3.0
requests==2.25.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
retrying==1.3.3
rsa==4.7; python_version >= '3.6'
sacremoses==0.0.43
send2trash==1.5.0
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.2.0; python_version >= '3.5'
starlette==0.13.6; python_version >= '3.6'
tensorboard-plugin-wit==1.8.0
tensorboard==2.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1'
tensorflow-datasets==4.2.0
tensorflow-estimator==2.4.0
tensorflow-metadata==0.27.0; python_version >= '3.6' and python_version < '4'
tensorflow-gpu
termcolor==1.1.0
terminado==0.9.2; python_version >= '3.6'
testpath==0.4.4
tokenizers==0.9.4
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
tornado==6.1; python_version >= '3.5'
tqdm==4.49.0; python_version >= '2.6' and python_version not in '3.0, 3.1'
traitlets==5.0.5; python_version >= '3.7'
transformers==4.2.2
typed-ast==1.4.2; implementation_name == 'cpython' and python_version < '3.8'
typing-extensions==3.7.4.3; python_version < '3.8' and python_version < '3.8'
urllib3==1.26.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uvicorn==0.13.3
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==1.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
wheel==0.36.2; python_version >= '3'
widgetsnbextension==3.5.1
wrapt==1.12.1
xxhash==2.0.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
yapf==0.30.0
zipp==3.4.0; python_version < '3.8'
